+++ src/boxes/tfhd.js
@@ -29,7 +29,7 @@ export default {
     ret.track_ID = r.bytesToInt(4);
 
     if (hasBaseDataOffset) {
-      ret.base_data_offset = r.bytesToUint64();
+      ret.base_data_offset = r.bytesToUint64BigInt();
     }
     if (hasSampleDescriptionIndex) {
       ret.sample_description_index = r.bytesToInt(4);