+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 - Minimum supported version is now es2017
+- 64-bit fields are now `BigInt`, not JS `number`, to be able to express all possible values without loss
 
 ### Features